Remove ActivityManager
